cdc: move package tp to internal
