+++ cdc/scheduler/internal/tp/agent.go
@@ -17,15 +17,15 @@ import (
 	"context"
 
 	"github.com/pingcap/tiflow/cdc/model"
-	"github.com/pingcap/tiflow/cdc/scheduler"
-	"github.com/pingcap/tiflow/cdc/scheduler/tp/schedulepb"
+	"github.com/pingcap/tiflow/cdc/scheduler/internal"
+	"github.com/pingcap/tiflow/cdc/scheduler/internal/tp/schedulepb"
 )
 
-var _ scheduler.Agent = (*agent)(nil)
+var _ internal.Agent = (*agent)(nil)
 
 type agent struct {
 	trans     transport
-	tableExec scheduler.TableExecutor
+	tableExec internal.TableExecutor
 
 	tables       map[model.TableID]*schedulepb.TableStatus
 	runningTasks map[model.TableID]*schedulepb.Message
@@ -36,7 +36,7 @@ func (a *agent) Tick(ctx context.Context) error {
 }
 
 func (a *agent) GetLastSentCheckpointTs() (checkpointTs model.Ts) {
-	return scheduler.CheckpointCannotProceed
+	return internal.CheckpointCannotProceed
 }
 
 func (a *agent) Close() error {
